More cleanup of unused functions. Moved a few things that are only used in tests out to a test file
