+++ WEB-INF/classes/gov/noaa/pfel/coastwatch/sgt/SgtMap.java
@@ -296,87 +296,6 @@ public class SgtMap {
   }
 
   /**
-   * This is an alternative version of makeMap which just plots grid data. The parameters match the
-   * same-named parameters for the main makeMap.
-   */
-  public static List<PrimitiveArray> makeMap(
-      int legendPosition,
-      String legendTitle1,
-      String legendTitle2,
-      String imageDir,
-      String logoImageFile,
-      double minX,
-      double maxX,
-      double minY,
-      double maxY,
-      String drawLandMask,
-      boolean plotGridData,
-      Grid gridGrid,
-      double gridScaleFactor,
-      double gridAltScaleFactor,
-      double gridAltOffset,
-      String gridPaletteFileName,
-      String gridBoldTitle,
-      String gridTitle2,
-      String gridTitle3,
-      String gridTitle4,
-      int drawLakesAndRivers,
-      Graphics2D g2,
-      int baseULXPixel,
-      int baseULYPixel,
-      int imageWidthPixels,
-      int imageHeightPixels,
-      int boundaryResAdjust,
-      double fontScale
-      // , String customFileName
-      ) throws Exception {
-
-    return makeMap(
-        false,
-        legendPosition,
-        legendTitle1,
-        legendTitle2,
-        imageDir,
-        logoImageFile,
-        minX,
-        maxX,
-        minY,
-        maxY,
-        drawLandMask,
-        plotGridData,
-        gridGrid,
-        gridScaleFactor,
-        gridAltScaleFactor,
-        gridAltOffset,
-        gridPaletteFileName,
-        gridBoldTitle,
-        gridTitle2,
-        gridTitle3,
-        gridTitle4,
-        drawLakesAndRivers,
-        false, // plotContourData,
-        null, // contourGrid,
-        1,
-        1,
-        0, // double contourScaleFactor, contourAltScaleFactor, contourAltOffset,
-        "10, 14", // contourDrawLinesAt,
-        new Color(0x990099), // contourColor
-        "Contour Bold Title",
-        "cUnits",
-        "Contour Title2 and more text",
-        "2004-01-05 to 2004-01-0C", // contourDateTime,
-        "Data courtesy of blah blah blah", // Contour data
-        new ArrayList<>(), // graphDataLayers
-        g2,
-        baseULXPixel,
-        baseULYPixel,
-        imageWidthPixels,
-        imageHeightPixels,
-        boundaryResAdjust,
-        fontScale);
-  }
-
-  /**
    * This uses SgtMap to plot data on a map. Strings should be "" if not needed. This is not static
    * because it uses boundary.
    *
