+++ src/archive/gov/noaa/pfel/coastwatch/util/SSR.java
@@ -473,5 +473,33 @@ System.out.println(
       return null;
     }
   }
+
+  /**
+   * This gets the bytes from a file.
+   *
+   * @param fileName If compressed file, this reads the decompressed, first file in the archive.
+   * @return a byte[] with the response.
+   * @throws Exception if error occurs
+   */
+  public static byte[] getFileBytes(String fileName) throws Exception {
+    try (InputStream is = File2.getDecompressedBufferedInputStream(fileName)) {
+      long time = System.currentTimeMillis();
+      byte buffer[] = new byte[1024];
+      ByteArrayOutputStream baos = new ByteArrayOutputStream();
+      for (int s; (s = is.read(buffer)) != -1; ) baos.write(buffer, 0, s);
+      is.close();
+      if (reallyVerbose)
+        String2.log(
+            "  SSR.getFileBytes "
+                + fileName
+                + " finished. TIME="
+                + (System.currentTimeMillis() - time)
+                + "ms");
+      return baos.toByteArray();
+    } catch (Exception e) {
+      // String2.log(e.toString());
+      throw new Exception("ERROR while reading file=" + fileName + " : " + e, e);
+    }
+  }
     
 }
