+++ src/test/java/gov/noaa/pfel/erddap/ArchiveADatasetTests.java
@@ -47,7 +47,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     String today = Calendar2.getCurrentISODateTimeStringZulu().substring(0, 10);
@@ -110,7 +110,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     // decompress and look at contents
@@ -213,7 +213,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     // decompress and look at contents
@@ -315,7 +315,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     String today = Calendar2.getCurrentISODateTimeStringZulu().substring(0, 10);
@@ -385,7 +385,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     // decompress and look at contents
@@ -485,7 +485,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     String today = Calendar2.getCurrentISODateTimeStringZulu().substring(0, 10);
@@ -552,7 +552,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     // decompress and look at contents
@@ -651,7 +651,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     String today = Calendar2.getCurrentISODateTimeStringZulu().substring(0, 10);
@@ -712,7 +712,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(5000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(5000);
 
     // decompress and look at contents
@@ -809,7 +809,7 @@ class ArchiveADatasetTests {
 
     // display it (in 7zip)
     Math2.sleep(10000);
-    // Test.displayInBrowser("file://" + targzName);
+    // TestUtil.displayInBrowser("file://" + targzName);
     // Math2.sleep(10000);
 
     // decompress and look at contents
