+++ src/test/java/gov/noaa/pfel/erddap/dataset/EDDGridFromAudioFilesTests.java
@@ -539,7 +539,7 @@ class EDDGridFromAudioFilesTests {
             + "9.375E-5,-7458\n"
             + "...\n";
     Test.ensureEqual(results, expected, "\nresults=\n" + results);
-    // Test.displayInBrowser("file://" + dir + tName);
+    // TestUtil.displayInBrowser("file://" + dir + tName);
     // String2.pressEnterToContinue("Close audio player when done.");
     File2.delete(dir + tName);
