+++ src/test/java/gov/noaa/pfel/erddap/dataset/EDDTableFromAudioFilesTests.java
@@ -383,7 +383,7 @@ class EDDTableFromAudioFilesTests {
             + "9.375E-5,-7458\n"
             + "...\n";
     Test.ensureEqual(results, expected, "\nresults=\n" + results);
-    // Test.displayInBrowser("file://" + dir + tName);
+    // TestUtil.displayInBrowser("file://" + dir + tName);
     // String2.pressEnterToContinue("Close audio player when done.");
     File2.delete(dir + tName);
   }
